requirements.txt
Flask
flask-mysql-connector
Flask-JWT-Extended
Flask-Mail
python-dotenv
bcrypt
twilio
